requirements.txt
requests==2.28.1
pycryptodome
bs4
grequests~=0.6.0
urllib3
gevent
gmpy2
libnum
cryptography
pillow
pip
tinyec
flask
ddddocr
